[refactor]: udp peer discovery
1. set the interface for peer discovery;
2. rename previous peer discovery class to UDP-specfic;
